refactor: package auth module and extract token parsing
